style: u16

+++ crates/cpu_sm83/src/alu/add.rs
@@ -18,8 +18,8 @@ pub(crate) fn alu_add_16(lhs: u16, rhs: u16) -> (u16, bool, bool) {
 
 pub(crate) fn alu_add_sp_r8(lhs: u16, rhs: i8) -> (u16, bool, bool) {
     let ret = lhs.wrapping_add_signed(rhs as i16);
-    let h = (lhs & 0xF) + (rhs as u16 & 0xF) > 0xF;
-    let c = (lhs & 0xFF) as i16 + rhs as u8 as i16 > 0xFF;
+    let h = (lhs & 0xF) + (rhs as u8 as u16 & 0xF) > 0xF;
+    let c = (lhs & 0xFF) + rhs as u8 as u16 > 0xFF;
 
     (ret, h, c)
 }
